change template, add name space, change form of elements management
