+++ app/models/element.rb
@@ -1,7 +1,7 @@
 class Element < ApplicationRecord
   enum selector_type: {
-    default: 0,
     html_id: 1,
-    html_name: 2
+    html_name: 2,
+    xpath: 3
   }
 end
