+++ app/views/admin/elements/index.slim
@@ -9,4 +9,5 @@ table.table.table-striped
 
   tbody
     = render @elements
-  
+
+= link_to "Add new element", new_admin_element_path, class: "btn btn-primary"
