--- app/views/elements/_element.slim
@@ -1,6 +1,0 @@
-tr
-  td = element.name
-  td = element.selector_type
-  td = element.selector_value
-  td = link_to "Edit", edit_element_path(element)
-  td = link_to "Delete", element, method: :delete, data: {confirm: "You sure?"}
