--- app/views/elements/_form.slim
@@ -1,13 +1,0 @@
-= form_for @element do |f|
-  div.form-group
-    = f.label "Name"
-    = f.text_field :name, id: "element-name", class: "form-control"
-  div.form-group
-    = f.label "Selector type"
-    = f.select :selector_type, Element.selector_types.map{|k, _| [k, k]}.to_h,
-      id: "element-selector-type", class: "form-control"
-  div.form-group
-    = f.label "Selector value"
-    = f.text_field :selector_value, id: "element-selector-value",
-      class: "form-control"
-  = f.submit "Submit", class: "btn btn-primary"
